added code to handle API request exceptions

+++ app/Http/Controllers/UserController.php
@@ -25,8 +25,10 @@ class UserController extends Controller
 
     public function show($user_id): \Illuminate\Http\JsonResponse
     {
-
-        $user = User::retrieveRecordById($user_id);
+        $user = User::find($user_id);
+        if(is_null($user)){
+            return response()->json('User record not found', 404);
+        }
         $addresses = (new \App\Models\Address)->retrieveRecordsById($user_id);
 
         return response()->json([$user, $addresses], 200);
@@ -40,8 +42,12 @@ class UserController extends Controller
 
     }
 
-    public function updateUser(StoreUserPutRequest $request, User $user): \Illuminate\Http\JsonResponse
+    public function updateUser(StoreUserPutRequest $request, $user_id): \Illuminate\Http\JsonResponse
     {
+        $user = User::find($user_id);
+        if(is_null($user)){
+            return response()->json('User record not found', 404);
+        }
         $user->update($request->all());
         return response()->json($user, 200);
 
@@ -54,19 +60,31 @@ class UserController extends Controller
 
     }
 
-
-    //ISSUE
-    public function updateAddress(StorePutRequest $request, Address $address): \Illuminate\Http\JsonResponse
+    public function updateAddress(StorePutRequest $request, $address_id): \Illuminate\Http\JsonResponse
     {
+        $user_id = $request->input('user_id');
+        $user = User::find($user_id);
+        $address = Address::find($address_id);
+        if(is_null($address)){
+            return response()->json('Address record not found', 404);
+        }
+        if(is_null($user)){
+            return response()->json('User with specified ID does not exist', 404);
+        }
         $address->update($request->all());
         return response()->json($address, 200);
 
+
     }
 
-    public function deleteAddress(Request $request, Address $address): \Illuminate\Http\JsonResponse
+    public function deleteAddress(Request $request, $address_id): \Illuminate\Http\JsonResponse
     {
+        $address = Address::find($address_id);
+        if(is_null($address)){
+            return response()->json('Address record not found', 404);
+        }
         $address->delete();
-        return response()->json(null, 204);
+        return response()->json('record has been deleted', 204);
 
     }
